update path (80%)

+++ SVG_Reader/src/Path.cpp
@@ -163,15 +163,18 @@ Path::Path(const Path& other) : Figure(other) {
 
 // Calculate method 
 float Path::computeBinominal(int n, int k) {
-	float value = 1.0;
+	double value = 1.0;
 
-	for (int i = 1; i <= k; i++) {
+	for (int i = 1; i <= k; i++)
+	{
+
 		value = value * ((n + 1 - i) / i);
 	}
 
 	if (n == k) {
 		value = 1;
 	}
+
 	return value;
 }
 
@@ -179,15 +182,18 @@ vector<Point> Path::CVertices(vector<Point> Position) {
 	vector<Point> CurvePositions;
 
 	int n = Position.size() - 1;
-	for (float t = 0.0; t <= 1.0; t += 0.001) {
+
+	for (double t = 0.0; t <= 1.0; t += 0.001)
+	{
 		Point CurvePos = { 0.0, 0.0 };
+
 		for (int i = 0; i <= n; ++i) {
 			CurvePos.x += computeBinominal(n, i) * pow((1 - t), (n - i)) * pow(t, i) * Position[i].x;
 			CurvePos.y += computeBinominal(n, i) * pow((1 - t), (n - i)) * pow(t, i) * Position[i].y;
 		}
+
 		CurvePositions.push_back(CurvePos);
 	}
-
 	return CurvePositions;
 }
